+++ app/controllers/application_controller.rb
@@ -3,6 +3,7 @@ class ApplicationController < ActionController::Base
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
+  before_action :set_paper_trail_whodunnit
 
   # Globally rescue Authorization Errors in controller.
   #rescue_from Pundit::NotAuthorizedError, with: :user_not_authorized
@@ -13,15 +14,16 @@ class ApplicationController < ActionController::Base
 
   before_action :configure_permitted_parameters, if: :devise_controller?
 
+  def switch_chapter
+    session[:current_chapter_id] = current_chapter == Chapter::BASTION ? Chapter::HOLURHEIM.id : Chapter::BASTION.id
+    redirect_to :back
+  end
+
   protected
 
   def configure_permitted_parameters
     devise_parameter_sanitizer.for(:sign_up) << :name
     devise_parameter_sanitizer.for(:account_update) << :name
-  end
-
-  def user_for_paper_trail
-    current_user ? current_user.try(:id) : User.find(1)
   end
 
   def pundit_user