+++ app/models/character_skill.rb
@@ -1,5 +1,4 @@
-class CharacterSkill < ActiveRecord::Base
-  has_paper_trail
+class CharacterSkill < ApplicationRecord
   belongs_to :character, inverse_of: :character_skills
   belongs_to :skill, inverse_of: :character_skills