+++ app/models/origin.rb
@@ -1,9 +1,21 @@
-class Origin < ActiveRecord::Base
-  has_paper_trail
+class Origin < ApplicationRecord
   has_many :characters, through: :character_origins, inverse_of: :origins
   has_many :character_origins, inverse_of: :origin
 
-  validates :source, inclusion: {in: (Character::RACES|Character::CULTURES) }
+  SOURCES = (Character::RACES|Character::CULTURES|[
+    "Template: Clocksmith",
+    "Template: Proto"]
+  )
+  SAMPLE = [
+    "Birthright: Arcane Aptitude",
+    "Birthright: Gnomish Luck",
+    "Birthright: Proto Form",
+    "Birthright: Proto Revelation",
+    "Origin: Cryogen",
+    "Origin: Common House of Arcane",
+  ]
+
+  validates :source, inclusion: {in: SOURCES }
   validates :name, presence: true
 
   def display_name