+++ app/models/temporary_effect.rb
@@ -1,5 +1,4 @@
-class TemporaryEffect < ActiveRecord::Base
-  has_paper_trail
+class TemporaryEffect < ApplicationRecord
   belongs_to :character
 
   validates_presence_of :attr