+++ config/application.rb
@@ -23,5 +23,6 @@ module SimRep
     # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
     # config.i18n.default_locale = :de
     config.active_record.raise_in_transactional_callbacks = true
+    config.web_console.development_only = false # srrrrrryyyyyy
   end
 end