+++ db/migrate/20150725144615_add_default_timeunits_to_npc_shifts.rb
@@ -1,5 +1,5 @@
 class AddDefaultTimeunitsToNpcShifts < ActiveRecord::Migration
-  class NpcShift < ActiveRecord::Base; end
+  class NpcShift < ApplicationRecord; end
 
   def up
     change_column_default :npc_shifts, :hours_to_money, 0